ref: move drop const to utils
