+++ components/carousel/CarouselTypeGenerative.vue
@@ -9,10 +9,6 @@
 
 <script lang="ts" setup>
 import { useCarouselGenerativeNftEvents } from './utils/useCarouselEvents'
-import {
-  AHK_GENERATIVE_DROPS,
-  AHP_GENERATIVE_DROPS,
-} from '@/components/collection/drop/const'
 
 const { nfts, ids } = await useCarouselGenerativeNftEvents(
   AHK_GENERATIVE_DROPS,
