+++ components/collection/drop/const.ts
@@ -4,19 +4,3 @@ const now = new Date()
 export const countDownTime = endOfWeek(now).getTime()
 
 export const MINT_ADDRESS = '15CoYMEnJhhWHvdEPXDuTBnZKXwrJzMQdcMwcHGsVx5kXYvW'
-
-export const DEFAULT_DROP = '/ahp/drops/wallstreet'
-
-export const AHK_GENERATIVE_DROPS = [
-  '176', // Chained
-]
-
-export const AHP_GENERATIVE_DROPS = [
-  '50', // .motherboard
-  '49', // wallstreet
-  '46', // Snowflakes
-  '40', // Swirls
-  '38', // Generativ Art - Pare1d0scope
-]
-
-export const AHP_POPULAR_DROP_COLLECTIONS = [...AHP_GENERATIVE_DROPS]
